deployment/requirements.txt
torch
mlflow
seldon-core
boto3
pillow
numpy
